Add Git hooks, update .gitignore
Added Git hooks to control .ipynb cells. Pre-commit hook removes
plumbing cells that are only necessary for conversion to html for
display on the website. Post-commit hook restores the removed cells
in preparation for uploading on personal website. Pre-push validates
that no keywords are present that indicate the file is not yet ready
for upload. .gitignore updated with new comments and reflects the new
usage of symlinks in the project.